refactor: logger under msg and don't use logger directly in mutex
